+++ embed/example-react/src/main.tsx
@@ -70,84 +70,11 @@ const App = observer(() => {
     ctx.call.startCall(inputRef.current.value, undefined, true)
   }
 
-  const updateDontShowIfFocusing = (value: boolean) => {
-    phone.setDontShowNotificationIfFocusing(value)
-    loadNotificationOptions()
-  }
-
-  const updateCloseAllOnFocus = (value: boolean) => {
-    phone.setCloseAllNotificationOnFocus(value)
-    loadNotificationOptions()
-  }
-
-  const updateCloseOnAnswer = (value: boolean) => {
-    phone.setCloseNotificationOnCallAnswer(value)
-    loadNotificationOptions()
-  }
-
-  const updateCloseOnEnd = (value: boolean) => {
-    phone.setCloseNotificationOnCallEnd(value)
-    loadNotificationOptions()
-  }
-
-  const updateDisableNotification = (value: boolean) => {
-    phone.setDisableNotification(value)
-    loadNotificationOptions()
-  }
-
-  const updateInterval = (value: number) => {
-    phone.setNotificationInterval(value)
-    loadNotificationOptions()
-  }
-
-  const loadNotificationOptions = () => {
-    const opts = phone.getNotificationOptions()
-    if (opts) {
-      setNotificationOptions({ ...opts })
-    }
-  }
-
-  useEffect(() => {
-    loadNotificationOptions()
-  }, [])
-
   return (
     <div className='app'>
       <span>Web Phone - {version.webphone} | </span>
       <span>JsSIP - {version.jssip} | </span>
       <span>{version.bundleIdentifier} </span>
-      <hr />
-
-      <h3>Notification Options</h3>
-      <div>
-        <span>dontShowNotificationIfFocusing: </span>
-        <span>
-          {notificationOptions?.dontShowNotificationIfFocusing?.toString()}
-        </span>
-        <button onClick={() => updateDontShowIfFocusing(true)}>true</button>
-        <button onClick={() => updateDontShowIfFocusing(false)}>false</button>
-      </div>
-      <hr />
-      <div>
-        <span>disableNotification: </span>
-        <span>{notificationOptions?.disableNotification?.toString()}</span>
-        <button onClick={() => updateDisableNotification(true)}>true</button>
-        <button onClick={() => updateDisableNotification(false)}>false</button>
-      </div>
-      <hr />
-
-      <div>
-        <span>notificationInterval (ms): </span>
-        <span>{notificationOptions?.notificationInterval}</span>
-        <input ref={displayIntervalRef} type='number' />
-        <button
-          onClick={() => {
-            updateInterval(Number(displayIntervalRef.current.value))
-          }}
-        >
-          Update
-        </button>
-      </div>
       <hr />
 
       <span>Status: </span>