+++ src/stores/addCallHistory.ts
@@ -5,30 +5,32 @@ import { Notifications } from 'react-native-notifications'
 import { v4 as newUuid } from 'uuid'
 
 import { isAndroid, isIos } from '#/config'
+import { embedApi } from '#/embed/embedApi'
 import { isEmbed } from '#/embed/polyfill'
 import { Call } from '#/stores/Call'
 import { getPartyName, getPartyNameAsync } from '#/stores/contactStore'
 import { ctx } from '#/stores/ctx'
 import { intl } from '#/stores/intl'
 import { BrekekeUtils, CallLogType } from '#/utils/BrekekeUtils'
+import { jsonStable } from '#/utils/jsonStable'
 import { permForCallLog } from '#/utils/permissions'
 import type { ParsedPn } from '#/utils/PushNotification-parse'
 import { waitTimeout } from '#/utils/waitTimeout'
 import { webShowNotification } from '#/utils/webShowNotification'
 
 let alreadyAddHistoryMap: { [pnId: string]: true } = {}
-export const parseReasonCancelCall = (reason?: string) => {
+
+const parseReasonCancelCall = (reason?: string) => {
   if (!reason) {
     return
   }
-
   const m = reason.match(/"([^"]+)"/i)
   if (!m) {
     return
   }
   return m[1]
 }
-export const getUserInfoFromReasons = (reason?: string | false) => {
+const getUserInfoFromReasons = (reason?: string | false) => {
   if (!reason) {
     return
   }
@@ -70,7 +72,8 @@ export const addCallHistory = async (
     // voice mail
     return
   }
-  const pnId = isTypeCall ? c.pnId : c.id
+
+  const pnId = (isTypeCall ? c.pnId : c.id) || c.id || jsonStable(c)
   if (pnId) {
     if (alreadyAddHistoryMap[pnId]) {
       return
@@ -264,7 +267,12 @@ const presentNotification = async (c: CallHistoryInfo) => {
   }
   const title = intl`Missed call`
   const body = await getBodyForNotification(c)
-  if (isEmbed) {
+  // show notification call completed elsewhere in web embed api
+  if (
+    isEmbed &&
+    c.answeredBy &&
+    embedApi._notificationOptions?.notificationCallCompletedElseWhere
+  ) {
     webShowNotification({
       type: 'call',
       id: c.id,